+++ Communication/Packets/Incoming/Moderation/CallForHelpPendingCallsDeletedEvent.cs
@@ -6,30 +6,12 @@ namespace Plus.Communication.Packets.Incoming.Moderation;
 
 internal class CallForHelpPendingCallsDeletedEvent : IPacketEvent
 {
-    private readonly IModerationManager _moderationManager;
-    private readonly IGameClientManager _clientManager;
+    private readonly IModerationTicketService _moderationTicketService;
 
-    public CallForHelpPendingCallsDeletedEvent(IModerationManager moderationManager, IGameClientManager clientManager)
+    public CallForHelpPendingCallsDeletedEvent(IModerationTicketService moderationTicketService)
     {
-        _moderationManager = moderationManager;
-        _clientManager = clientManager;
+        _moderationTicketService = moderationTicketService;
     }
 
-    public Task Parse(GameClient session, IIncomingPacket packet)
-    {
-        var habbo = session.GetHabbo();
-        if (habbo == null)
-            return Task.CompletedTask;
-
-        if (_moderationManager.UserHasTickets(habbo.Id))
-        {
-            var pendingTicket = _moderationManager.GetTicketBySenderId(habbo.Id);
-            if (pendingTicket != null)
-            {
-                pendingTicket.Answered = true;
-                _clientManager.SendPacket(new ModeratorSupportTicketComposer(habbo.Id, pendingTicket), "mod_tool");
-            }
-        }
-        return Task.CompletedTask;
-    }
+    public Task Parse(GameClient session, IIncomingPacket packet) => _moderationTicketService.DeletePendingCalls(session);
 }